Fix unsigned comparison.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk/net-snmp@4562 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/snmp_api.c
@@ -4962,7 +4962,7 @@ snmp_add_var(struct snmp_pdu *pdu,
 	      }
 	    }
 	    ix = ltmp/8;
-	    if (ix >= tint) tint = ix+1;
+	    if (ix >= (int)tint) tint = ix+1;
 	    bit = 0x80 >> ltmp%8;
 	    buf[ix] |= bit;
 	  }
